% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realmlp-package.R
\docType{package}
\name{realmlp-package}
\alias{realmlp}
\alias{realmlp-package}
\title{realmlp: RealMLP with R}
\description{
Implementation of RealMLP, currently supporting RealMLP-TD-S. This R implementation is based soley on the RealMLP-TD-S standalone implementation (via dholzmueller/realmlp-td-s_standalone on github). RealMLP-TD-S stands for RealMLP with Tuned Defaults Simplified.
}
\author{
\strong{Maintainer}: Frank Hull \email{frankiethull@gmail.com}

}
\keyword{internal}
